Fixed race condition in check for existing backup.
Fixed sql statement to check for existing backup to only look for
backups after midnight of today. Before it was prone to a race condition
when the previous backup had finished a little bit later than 24 hours
ago. Not the behavoir should be the same as with SESSION_TYPE "TXT".